requirements.txt
Django==2.2.2
line-bot-sdk==1.8.0
gunicorn
django-heroku==0.3.1
psycopg2==2.8.3
pandas
numpy
scipy